Update dependency versions and fix minor issues
This update includes an upgrade to the `flutter_lints` package from version `1.0.0` to `4.0.0` in both the main and example apps. The 'loading_animation_widget' dependency has been removed, as well as Cupertino Icons package and lints package updated. Corrected a typo mistake in the 'detail_exception_handler' message and changed 'UserModel' to a final class.